test: fix distance_m type, add busy and reminder upper-bound tests
- Cast distance_m to float8 in findNearbyDrivers SQL so pg returns a
  proper JS number instead of a string; removes Number() workarounds
  from three test assertions
- Add missing 'excludes busy drivers' test to location.test.ts
- Add missing upper-bound reminder window test to jobs.test.ts

+++ src/services/location.ts
@@ -39,7 +39,7 @@ export async function findNearbyDrivers(
        ROUND(ST_Distance(
          d.location,
          ST_MakePoint($1, $2)::geography
-       )::numeric, 0) AS distance_m
+       )::numeric, 0)::float8 AS distance_m
      FROM drivers d
      WHERE d.status = 'available'
        AND d.is_approved = true
